Add files via upload
This is a sample code for ordinary least squares model to find the optimize parameters and find the best hypothesis for the regression model to optimize the squared loss. 
The code also provides the ridge regression model to prevent the linear regression model from overfitting by penalizing the parameters to smaller values.